Finish RestControllerToInsertOnePost

+++ src/main/java/com/mycompany/myapp/domain/PostVO.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class PostVO {
 
-	private int id;
-	private int userId;
+	private Long id;
+	private Long userId;
 	private String title;
 	private String content;
 	private String createdAt;
@@ -15,19 +15,19 @@ public class PostVO {
 		super();
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
@@ -57,8 +57,8 @@ public class PostVO {
 
 	@Override
 	public String toString() {
-		String info = "PostVO: id: " + id + ", userId: " + userId + ", title: " + title + ", content: "
-				+ content + "createdAt: " + createdAt;
+		String info = "PostVO: id: " + id + ", userId: " + userId + ", title: " + title + ", content: " + content
+				+ "createdAt: " + createdAt;
 		return info;
 	}
